+++ hazelcast/src/main/java/com/hazelcast/management/MapConfigAdapter.java
@@ -93,7 +93,7 @@ public class MapConfigAdapter implements JsonSerializable, DataSerializable {
         config.setMinEvictionCheckMillis(in.readLong());
         config.setTimeToLiveSeconds(in.readInt());
         config.setMaxIdleSeconds(in.readInt());
-        config.setMaxSizeConfig(new MaxSizeConfig().setSize(in.readInt())
+        config.setMaxSizeConfig(new MaxSizeConfig().setSize(in.readLong())
                 .setMaxSizePolicy(MaxSizeConfig.MaxSizePolicy.valueOf(in.readUTF())));
         config.setReadBackupData(in.readBoolean());
         config.setEvictionPolicy(EvictionPolicy.valueOf(in.readUTF()));
@@ -110,7 +110,7 @@ public class MapConfigAdapter implements JsonSerializable, DataSerializable {
         out.writeLong(config.getMinEvictionCheckMillis());
         out.writeInt(config.getTimeToLiveSeconds());
         out.writeInt(config.getMaxIdleSeconds());
-        out.writeInt(config.getMaxSizeConfig().getSize());
+        out.writeLong(config.getMaxSizeConfig().getSize());
         out.writeUTF(config.getMaxSizeConfig().getMaxSizePolicy().toString());
         out.writeBoolean(config.isReadBackupData());
         out.writeUTF(config.getEvictionPolicy().name());
